refactor(pattern): 抽出util文件
