+++ src/utils/pattern/dot.ts
@@ -1,51 +1,6 @@
-import { DotPatternCfg, SquarePatternCfg, LinePatternCfg } from '../../types/pattern';
+import { DotPatternCfg } from '../../types/pattern';
 import { deepAssign } from '../../utils';
-
-/**
- * 初始化 cavnas，设置宽高等
- */
-export function initCanvas(width: number, height: number = width): HTMLCanvasElement {
-  const canvas = document.createElement('canvas');
-  // ~~~ 后续再行测试 ~~~
-  // const pixelRatio = window?.devicePixelRatio || 2;
-  // const logicalWidth = width;
-  // const logicalHeight = height;
-  // 画布尺寸
-  // canvas.width = logicalWidth * pixelRatio;
-  // canvas.height = logicalHeight * pixelRatio;
-  // 显示尺寸
-  // canvas.style.width = `${logicalWidth}px`;
-  // canvas.style.height = `${logicalHeight}px`;
-  // ~~~ 后续再行测试 ~~~
-  canvas.width = width;
-  canvas.height = height;
-
-  return canvas;
-}
-
-/**
- * 绘制背景
- *
- * @param context
- * @param cfg
- * @param width
- * @param height
- */
-export function drawBackground(
-  context: CanvasRenderingContext2D,
-  cfg: DotPatternCfg | LinePatternCfg | SquarePatternCfg,
-  width: number,
-  height: number = width
-) {
-  const { backgroundColor, opacity } = cfg;
-
-  context.globalAlpha = opacity;
-  context.fillStyle = backgroundColor;
-
-  context.beginPath();
-  context.fillRect(0, 0, width, height);
-  context.closePath();
-}
+import { getUnitPatternSize, initCanvas, drawBackground, getSymbolsPosition } from './util';
 
 /**
  * dotPattern的默认配置
@@ -62,20 +17,6 @@ export const defaultDotPatternCfg = {
   isStagger: true,
   mode: 'repeat',
 };
-
-/**
- * 计算贴图单元大小
- *
- * @param radius 圆点半径
- * @param padding 圆点间隔
- * @param isStagger 是否交错
- * @reutrn 返回贴图单元大小
- */
-export function getUnitPatternSize(radius: number, padding: number, isStagger: boolean): number {
-  // 如果交错, unitSize 放大两倍
-  const unitSize = radius * 2 + padding;
-  return isStagger ? unitSize * 2 : unitSize;
-}
 
 /**
  * 绘制圆点
@@ -102,24 +43,6 @@ export function drawDot(context: CanvasRenderingContext2D, cfg: DotPatternCfg, x
 }
 
 /**
- * 计算圆点坐标
- *
- * @param unitSize 贴图单元大小
- * @param isStagger 是否交错
- * @reutrn 圆点中心坐标 x,y 数组集合
- */
-export function getDots(unitSize: number, isStagger: boolean): number[][] {
-  // 如果交错, 交错绘制 dot
-  const dots = isStagger
-    ? [
-        [unitSize * (1 / 4), unitSize * (1 / 4)],
-        [unitSize * (3 / 4), unitSize * (3 / 4)],
-      ]
-    : [[unitSize * (1 / 2), unitSize * (1 / 2)]];
-  return dots;
-}
-
-/**
  * 创建 dot pattern，返回 HTMLCanvasElement
  *
  * @param cfg
@@ -131,8 +54,8 @@ export function createDotPattern(cfg?: DotPatternCfg): CanvasPattern {
   const { radius, padding, isStagger } = dotCfg;
 
   // 计算 画布大小，dots的位置
-  const unitSize = getUnitPatternSize(radius, padding, isStagger);
-  const dots = getDots(unitSize, isStagger);
+  const unitSize = getUnitPatternSize(radius * 2, padding, isStagger);
+  const dots = getSymbolsPosition(unitSize, isStagger);
 
   // 初始化 patternCanvas
   const canvas = initCanvas(unitSize, unitSize);
